+++ src/oms_sensemaking/geospatial/sensemakers/cotravel.py
@@ -6,15 +6,14 @@ import uuid
 from collections import defaultdict
 from dataclasses import dataclass, field
 from datetime import datetime, timedelta
-from uuid import UUID, uuid4
+from uuid import UUID
 
 from geoalchemy2.shape import to_shape
 from oms_sdk.generated.generated_graphql_client.client import (
-    CreateAttributeInput,
-    CreateNodeInput,
+    CreateActivityInput,
     CreateRelationshipInput,
 )
-from oms_sdk.generated.generated_graphql_client.enums import AttributeType, Confidence, ObjectTier
+from oms_sdk.generated.generated_graphql_client.enums import Confidence
 from shapely import LineString, MultiLineString
 from sqlalchemy import func, join, select
 
@@ -38,9 +37,9 @@ class CotravelType(enum.Enum):
         if cotravel_type == cls.potential_duplicate:
             return SETTINGS.potential_duplicate_relationship_name
         if cotravel_type == cls.cotravel:
-            return SETTINGS.cotravel_event_name
+            return SETTINGS.cotravel_activity_name
         if cotravel_type == cls.lag_lead:
-            return SETTINGS.lag_lead_event_name
+            return SETTINGS.lag_lead_activity_name
 
         raise ValueError("Invalid CotravelType")
 
@@ -128,7 +127,6 @@ class Cotravel(FindingBase):
     """Represents a Cotravel Event"""
 
     FINDING_TYPE: FindingType = field(init=False, default=FindingType.GEO_COTRAVEL)
-    cotravel_id: UUID = field(init=False, default_factory=uuid4)
     track1: Track
     track2: Track
     start_time: datetime
@@ -215,12 +213,10 @@ class CotravelSensemaker(Sensemaker):
         super().__init__()
         self.version = (1, 0, 0)
         self.config = {
-            "cotravel_event_node_attribute_iri": SETTINGS.cotravel_event_node_attribute_iri,
             "cotravel_relationship_iri": SETTINGS.cotravel_relationship_iri,
-            "cotravel_track_to_event_relation_name": SETTINGS.cotravel_track_to_event_relation_name,
-            "cotravel_event_node_iri": SETTINGS.cotravel_event_node_iri,
-            "cotravel_event_name": SETTINGS.cotravel_event_name,
-            "lag_lead_event_name": SETTINGS.lag_lead_event_name,
+            "cotravel_activity_iri": SETTINGS.cotravel_activity_iri,
+            "cotravel_activity_name": SETTINGS.cotravel_activity_name,
+            "lag_lead_activity_name": SETTINGS.lag_lead_activity_name,
             "geo_sensemaker_event_tag": SETTINGS.geo_sensemaker_event_tag,
         }
         self.oms_crud_tool = oms_crud_tool
@@ -518,62 +514,48 @@ class CotravelSensemaker(Sensemaker):
             self.version_string,
         ]
 
-        create_node_input = CreateNodeInput(
+        create_activity_input1 = CreateActivityInput(
             acm=cotravel.get_acm(),
+            tags=tags,
+            labels=labels,
+            classIri=SETTINGS.cotravel_activity_iri,
             name=name,
-            tier=ObjectTier.DERIVATIVE,
-            tags=[SETTINGS.geo_sensemaker_event_tag],
-            labels=labels,
-            classIri=SETTINGS.cotravel_event_node_iri,
-            ifcCodes=set(),
-            isNso=True,
-        )
-        published_node = self.oms_crud_tool.create_node(node_input=create_node_input)
-
-        # Save the created node id and type
-        cotravel.atoms_id = published_node.id
-        cotravel.atoms_type = AtomsType.NODE
-
-        # vehicle 1 relationship
-        create_relationship_input1 = CreateRelationshipInput(
-            tags=tags,
-            labels=labels,
-            name=f"{name} {SETTINGS.cotravel_track_to_event_relation_name}",
-            startNodeId=published_node.id,
-            endNodeId=cotravel.track1.node_id,
-            confidence=Confidence.HIGH,
-            acm=cotravel.get_acm(),
-            objectPropertyIri=SETTINGS.cotravel_relationship_iri,
+            state=SETTINGS.cotravel_activity_state,
             sourceId=source_id,
+            nodeId=cotravel.track1.node_id,
+            observationIds=cotravel.track1.observation_ids,
+            startTime=cotravel.start_time,
+            endTime=cotravel.last_time,
         )
-        # vehicle 2 relationship
-        create_relationship_input2 = CreateRelationshipInput(
+        create_activity_input2 = CreateActivityInput(
+            acm=cotravel.get_acm(),
             tags=tags,
             labels=labels,
-            name=f"{name} {SETTINGS.cotravel_track_to_event_relation_name}",
-            startNodeId=published_node.id,
+            classIri=SETTINGS.cotravel_activity_iri,
+            name=name,
+            state=SETTINGS.cotravel_activity_state,
+            sourceId=source_id,
+            nodeId=cotravel.track2.node_id,
+            observationIds=cotravel.track2.observation_ids,
+            startTime=cotravel.start_time,
+            endTime=cotravel.last_time,
+        )
+
+        self.oms_crud_tool.publish_activities([create_activity_input1, create_activity_input2])
+
+        create_relationship_input = CreateRelationshipInput(
+            tags=tags,
+            labels=labels,
+            name=f"{name} {SETTINGS.cotravel_relation_name}",
+            startNodeId=cotravel.track1.node_id,
             endNodeId=cotravel.track2.node_id,
-            confidence=Confidence.HIGH,
+            sourceId=source_id,
+            confidence=Confidence.UNKNOWN,
             acm=cotravel.get_acm(),
             objectPropertyIri=SETTINGS.cotravel_relationship_iri,
-            sourceId=source_id,
         )
-        self.oms_crud_tool.publish_relationships([create_relationship_input1, create_relationship_input2])
 
-        # attribute for geometry
-        created_attribute_input = CreateAttributeInput(
-            attributeIri=SETTINGS.cotravel_event_node_attribute_iri,
-            attributeValue="geo",
-            attributeDisplayValue="",
-            attributeType=AttributeType.GEOSPATIAL,
-            confidence=Confidence.HIGH,
-            tags=[SETTINGS.geo_sensemaker_event_tag],
-            labels=labels,
-            sourceId=source_id,
-            geometry=cotravel.to_geojson(),
-            nodeId=published_node.id,
-            acm=cotravel.get_acm(),
-            valueStart=cotravel.start_time,
-            valueEnd=cotravel.last_time,
-        )
-        self.oms_crud_tool.publish_attributes([created_attribute_input])
+        published_relationship = self.oms_crud_tool.create_relationship(create_relationship_input)
+
+        cotravel.atoms_id = published_relationship.id
+        cotravel.atoms_type = AtomsType.RELATIONSHIP