+++ CHANGELOG.md
@@ -7,8 +7,17 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 ### Changed
-
+- Loiters now return activities 
 ### Fixed
+
+## [0.19.6] - 2026-02-16
+
+### Added
+- Resolution Sensemaker now supports matching on any one of multiple criteria sets (e.g., `[BENUMBER+OSUFFX]` OR `[SK]`).
+
+### Changed
+- The Cotravel sensemaker now outputs two activities (for each "cotravelling" node) and one relationship (between the "cotravelling" nodes)
+- The Incursion Sensemaker no longer outputs/updates an attribute (only an activity).
 
 ## [0.19.5] - 2026-02-10